+++ private/plugins/mediagallery/include/lib-upload.php
@@ -32,7 +32,7 @@ function MG_videoThumbnail($aid, $srcImage, $media_filename ) {
         $ffmpeg_cmd = sprintf('"' . $_MG_CONF['ffmpeg_path'] . "/ffmpeg" . '" ' . $_MG_CONF['ffmpeg_command_args'],$srcImage, $thumbNail);
 
         $rc = UTL_execWrapper($ffmpeg_cmd);
-        COM_errorLog("MG Upload: FFMPEG returned: " . $rc);
+        Log::write('system',Log::DEBUG,"MG Upload: FFMPEG returned: " . $rc);
         if ( $rc != 1 ) {
             @unlink ($thumbNail);
             return 0;
@@ -52,7 +52,7 @@ function MG_processOriginal( $srcImage, $mimeExt, $mimeType, $aid, $baseFilename
     $newSrc = $srcImage;
 
     if ($_MG_CONF['verbose'] ) {
-        COM_errorLog("MG Upload: Entering MG_processOriginal()");
+        Log::write('system',Log::DEBUG,'MG Upload: Entering MG_processOriginal()');
     }
     $imgsize = @getimagesize("$srcImage");
     $imgwidth = $imgsize[0];
@@ -87,11 +87,11 @@ function MG_convertImage( $srcImage, $imageThumb, $imageDisplay, $mimeExt, $mime
     $makeSquare = 0;
 
     if ($_MG_CONF['verbose'] ) {
-        COM_errorLog("MG Upload: Entering MG_convertImage()");
+        Log::write('system',Log::DEBUG,'MG Upload: Entering MG_convertImage()');
     }
 
     if ($_MG_CONF['verbose'] ) {
-        COM_errorLog("MG Upload: Creating thumbnail image");
+        Log::write('system',Log::DEBUG,'MG Upload: Creating thumbnail image');
     }
     $imgsize = @getimagesize("$srcImage");
     if ( $imgsize == false &&
@@ -156,7 +156,7 @@ function MG_convertImage( $srcImage, $imageThumb, $imageDisplay, $mimeExt, $mime
         $tmpImage = $_MG_CONF['tmp_path'] . '/wip' . rand() . '.jpg';
         list($rc,$msg) = IMG_convertImageFormat($srcImage, $tmpImage, 'image/jpeg',0);
         if ( $rc == false ) {
-            COM_errorLog("MG_convertImage: Error converting uploaded image to jpeg format.");
+            Log::write('system',Log::ERROR,'MG_convertImage: Error converting uploaded image to jpeg format.');
             @unlink($srcImage);
             return array(false,$msg);
         }
@@ -173,7 +173,7 @@ function MG_convertImage( $srcImage, $imageThumb, $imageDisplay, $mimeExt, $mime
         }
     }
     if ( $rc == false ) {
-        COM_errorLog("MG_convertImage: Error resizing uploaded image to thumbnail size.");
+        Log::write('system',Log::ERROR,'MG_convertImage: Error resizing uploaded image to thumbnail size.');
         @unlink($srcImage);
         return array(false,$msg);
     }
@@ -326,7 +326,7 @@ function MG_processDir ($dir, $album_id, $purgefiles, $recurse ) {
            $file_extension = strtolower(substr(strrchr($filename,"."),1));
 
             if ( $MG_albums[$album_id]->max_filesize != 0 && filesize($filetmp) > $MG_albums[$album_id]->max_filesize) {
-                COM_errorLog("MG Upload: File " . $file . " exceeds maximum filesize for this album.");
+                Log::write('system',Log::ERROR,'MG Upload: File ' . $file . ' exceeds maximum filesize for this album.');
                 $statusMsg = sprintf($LANG_MG02['upload_exceeds_max_filesize'] . '<br/>',$file);
                 continue;
             }
@@ -435,13 +435,13 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
 
     $sizeofalbums = sizeof($MG_albums);
     if ($_MG_CONF['verbose']) {
-        COM_errorLog("MG Upload: *********** Beginning media upload process...");
-        COM_errorLog("Filename to process: " . $filename);
-        COM_errorLog("Size of MG_albums()=" . $sizeofalbums );
-        COM_errorLog("UID=" . $_USER['uid']);
-        COM_errorLog("album access=" . $MG_albums[$albums]->access );
-        COM_errorLog("album owner_id=" . $MG_albums[$albums]->owner_id );
-        COM_errorLog("member_uploads=" . $MG_albums[$albums]->member_uploads );
+        Log::write('system',Log::DEBUG,"MG Upload: *********** Beginning media upload process...");
+        Log::write('system',Log::DEBUG,"Filename to process: " . $filename);
+        Log::write('system',Log::DEBUG,"Size of MG_albums()=" . $sizeofalbums );
+        Log::write('system',Log::DEBUG,"UID=" . $_USER['uid']);
+        Log::write('system',Log::DEBUG,"album access=" . $MG_albums[$albums]->access );
+        Log::write('system',Log::DEBUG,"album owner_id=" . $MG_albums[$albums]->owner_id );
+        Log::write('system',Log::DEBUG,"member_uploads=" . $MG_albums[$albums]->member_uploads );
     }
 
     clearstatcache();
@@ -464,7 +464,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
     }
 
     if ( $MG_albums[$albums]->access != 3 && !$MG_albums[0]->owner_id && $MG_albums[$albums]->member_uploads == 0) {
-        COM_errorLog("Someone has tried to upload to an album in Media Gallery without the proper permissions.  User id: {$_USER['uid']}, Username: {$_USER['username']}, IP: " . $_SERVER['REMOTE_ADDR'],1);
+        Log::write('system',Log::ERROR,"Someone has tried to upload to an album in Media Gallery without the proper permissions.  User id: {$_USER['uid']}, Username: {$_USER['username']}, IP: " . $_SERVER['REMOTE_ADDR'],1);
         return array(false,$LANG_MG00['access_denied_msg']);
     }
 
@@ -498,8 +498,8 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
         $importSource = $filename;
     }
     if ( $rc != 1 ) {
-        COM_errorLog("Media Upload - Error moving uploaded file in generic processing....");
-        COM_errorLog("Media Upload - Unable to copy file to: " . $tmpPath);
+        Log::write('system',Log::ERROR,"Media Upload - Error moving uploaded file in generic processing....");
+        Log::write('system',Log::ERROR,"Media Upload - Unable to copy file to: " . $tmpPath);
         $errors++;
         $errMsg .= sprintf($LANG_MG02['move_error'],$filename);
         if ( $upload ) {
@@ -507,7 +507,6 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
         } else if ( !$purgefiles ) {
             @unlink($tmpPath);
         }
-        COM_errorLog("MG Upload: Problem uploading a media object");
         return array( false, $errMsg );
     }
 
@@ -533,7 +532,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
     $mimeInfo['type'] = $mimeExt;
 
     if ( !isset($mimeInfo['mime_type']) || $mimeInfo['mime_type'] == '' ) {
-        COM_errorLog("MG Upload: getID3 was unable to detect mime type - using PHP detection");
+        Log::write('system',Log::ERROR,"MG Upload: getID3 was unable to detect mime type - using PHP detection");
         $mimeInfo['mime_type'] = $filetype;
     }
 
@@ -544,7 +543,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
 	}
 
     if ($_MG_CONF['verbose']) {
-        COM_errorLog("MG Upload: found mime type of " . $mimeInfo['type']);
+        Log::write('system',Log::ERROR,"MG Upload: found mime type of " . $mimeInfo['type']);
     }
 
     if ( $mimeExt == '' || $mimeInfo['mime_type'] == 'application/octet-stream' || $mimeInfo['mime_type'] == '' ) {
@@ -646,7 +645,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
                 }
         }
         if ($_MG_CONF['verbose']) {
-            COM_errorLog("MG Upload: override mime type to: " . $mimeInfo['type'] . ' based upon file extension of: ' . $filetype );
+            Log::write('system',Log::DEBUG,"MG Upload: override mime type to: " . $mimeInfo['type'] . ' based upon file extension of: ' . $filetype );
         }
     }
 
@@ -726,7 +725,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
 
     $mimeType = $mimeInfo['mime_type'];
     if ( $_MG_CONF['verbose'] ) {
-        COM_errorLog("MG Upload: PHP detected mime type is : " . $filetype);
+        Log::write('system',Log::DEBUG,"MG Upload: PHP detected mime type is : " . $filetype);
     }
     if ( $filetype == 'video/x-m4v' ) {
         $mimeType = 'video/x-m4v';
@@ -800,11 +799,11 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
     $disp_media_filename = $media_filename . '.' . $mimeExt;
 
     if ( $_MG_CONF['verbose']) {
-        COM_errorLog("MG Upload: Stored filename is : " . $disp_media_filename);
+        Log::write('system',Log::DEBUG,"MG Upload: Stored filename is : " . $disp_media_filename);
     }
 
     if ( $_MG_CONF['verbose']) {
-        COM_errorLog("MG Upload: Mime Type: " . $mimeType);
+        Log::write('system',Log::DEBUG,"MG Upload: Mime Type: " . $mimeType);
     }
 
     switch ( $mimeType ) {
@@ -859,13 +858,13 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
             }
 
             if ( $_MG_CONF['verbose'] ) {
-                COM_errorLog("MG Upload: About to move/copy file");
+                Log::write('system',Log::DEBUG,"MG Upload: About to move/copy file");
             }
             $rc = @copy($filename, $media_orig);
 
             if ( $rc != 1 ) {
-                COM_errorLog("Media Upload - Error moving uploaded file....");
-                COM_errorLog("Media Upload - Unable to copy file to: " . $media_orig);
+                Log::write('system',Log::ERROR,"Media Upload - Error moving uploaded file....");
+                Log::write('system',Log::ERROR,"Media Upload - Unable to copy file to: " . $media_orig);
                 $errors++;
                 $errMsg .= sprintf($LANG_MG02['move_error'],$filename);
             } else {
@@ -875,7 +874,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
                 @chmod($media_orig, 0644);
 
                 if ( $_MG_CONF['verbose'] ) {
-                    COM_errorLog("MG Upload: Calling MG_convertImage()");
+                    Log::write('system',Log::DEBUG,"MG Upload: Calling MG_convertImage()");
                 }
 
                 // auto rotate
@@ -968,8 +967,8 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
 
             if ( $rc != 1 )
             {
-                COM_errorLog("MG Upload: Error moving uploaded file in video processing....");
-                COM_errorLog("Media Upload - Unable to copy file to: " . $media_orig);
+                Log::write('system',Log::ERROR,"MG Upload: Error moving uploaded file in video processing....");
+                Log::write('system',Log::ERROR,"Media Upload - Unable to copy file to: " . $media_orig);
                 $errors++;
                 $errMsg .= sprintf($LANG_MG02['move_error'],$filename);
             } else {
@@ -992,7 +991,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
 
             $rc = @copy($filename, $media_orig);
 
-    		COM_errorLog("MG Upload: Extracting audio meta data");
+    		Log::write('system',Log::DEBUG,"MG Upload: Extracting audio meta data");
 
         	if ( isset($mimeInfo['tags']['id3v1']['title'][0]) ) {
         		if ( $caption == '' ) {
@@ -1011,8 +1010,8 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
         	}
             if ( $rc != 1 )
             {
-                COM_errorLog("Media Upload - Error moving uploaded file in audio processing....");
-                COM_errorLog("Media Upload - Unable to copy file to: " . $media_orig);
+                Log::write('system',Log::ERROR,"Media Upload - Error moving uploaded file in audio processing....");
+                Log::write('system',Log::ERROR,"Media Upload - Unable to copy file to: " . $media_orig);
                 $errors++;
                 $errMsg .= sprintf($LANG_MG02['move_error'],$filename);
             } else {
@@ -1038,8 +1037,8 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
 
             if ( $rc != 1 )
             {
-                COM_errorLog("Media Upload - Error moving uploaded file in generic processing....");
-                COM_errorLog("Media Upload - Unable to copy file to: " . $media_orig);
+                Log::write('system',Log::ERROR,"Media Upload - Error moving uploaded file in generic processing....");
+                Log::write('system',Log::ERROR,"Media Upload - Unable to copy file to: " . $media_orig);
                 $errors++;
                 $errMsg .= sprintf($LANG_MG02['move_error'],$filename);
             } else {
@@ -1067,7 +1066,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
         if ( $upload ) {
             @unlink($tmpPath);
         }
-        COM_errorLog("MG Upload: Problem uploading a media object");
+        Log::write('system',Log::ERROR,"MG Upload: Problem uploading a media object");
         return array( false, $errMsg );
     }
 
@@ -1093,7 +1092,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
             $atttn = 1;
         }
         if ( $_MG_CONF['verbose']) {
-            COM_errorLog("MG Upload: Building SQL and preparing to enter database");
+            Log::write('system',Log::DEBUG,"MG Upload: Building SQL and preparing to enter database");
         }
 
         if ( $MG_albums[$albums]->enable_html != 1 ) {
@@ -1133,7 +1132,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
         }
 
         if ($_MG_CONF['verbose']) {
-            COM_errorLog("MG Upload: Inserting media record into mg_media");
+            Log::write('system',Log::DEBUG,"MG Upload: Inserting media record into mg_media");
         }
 
         $resolution_x = 0;
@@ -1224,7 +1223,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
 	        DB_query( $sql );
 
 	        if ( $_MG_CONF['verbose'] ) {
-	            COM_errorLog("MG Upload: Updating Album information");
+	            Log::write('system',Log::DEBUG,"MG Upload: Updating Album information");
 	        }
 	        $x = 0;
 	        $sql = "SELECT MAX(media_order) + 10 AS media_seq FROM " . $_TABLES['mg_media_albums'] . " WHERE album_id = " . $albums;
@@ -1267,7 +1266,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
 	                    $covername = $media_filename;
 	                }
 	                if ( $_MG_CONF['verbose']) {
-	                    COM_errorLog("MG Upload: Setting album cover filename to " . $covername);
+	                    Log::write('system',Log::DEBUG,"MG Upload: Setting album cover filename to " . $covername);
 	                }
 	                DB_query("UPDATE {$_TABLES['mg_albums']} SET album_cover_filename='" . $covername . "'" .
 	                         " WHERE album_id='" . $MG_albums[$albums]->id . "'");
@@ -1288,7 +1287,7 @@ function MG_getFile( $filename, $file, $albums, $caption = '', $description = ''
         MG_buildAlbumRSS( $albums );
         $c = glFusion\Cache::getInstance()->deleteItemsByTag('whatsnew');
     }
-    COM_errorLog("MG Upload: Successfully uploaded a media object");
+    Log::write('system',Log::INFO,"MG Upload: Successfully uploaded a media object");
 
     if ( $upload )
         @unlink($tmpPath);
@@ -1300,7 +1299,7 @@ function MG_attachThumbnail( $aid, $thumbnail, $mediaFilename, $origMediaFilenam
 
     $makeSquare = 0;
     if ($_MG_CONF['verbose']) {
-        COM_errorLog("MG Upload: Processing attached thumbnail: " . $thumbnail );
+        Log::write('system',Log::DEBUG,"MG Upload: Processing attached thumbnail: " . $thumbnail );
     }
 
     if ( $MG_albums[$aid]->tn_size == 3 || $MG_albums[$aid]->tn_size == 4 ) {
@@ -1348,7 +1347,7 @@ function MG_attachThumbnail( $aid, $thumbnail, $mediaFilename, $origMediaFilenam
             $tnExt = '.bmp';
             break;
         default:
-            COM_errorLog("MG_attachThumbnail: Invalid graphics type for attached thumbnail.");
+        Log::write('system',Log::ERROR,"MG_attachThumbnail: Invalid graphics type for attached thumbnail.");
             return false;
     }
     $attach_tn   = $mediaFilename . $tnExt;
@@ -1449,7 +1448,7 @@ function MG_notifyModerators( $aid ) {
             $row = DB_fetchArray($result);
             if ( $row['email'] != '' ) {
 			    if ($_MG_CONF['verbose'] ) {
-					COM_errorLog("MG Upload: Sending notification email to: " . $row['email'] . " - " . $row['username']);
+					Log::write('system',Log::DEBUG,"MG Upload: Sending notification email to: " . $row['email'] . " - " . $row['username']);
 				}
                 $toCount++;
                 $to[] = array('email' => $row['email'], 'name' => $row['username']);
@@ -1464,7 +1463,7 @@ function MG_notifyModerators( $aid ) {
             $msgData['to'] = $to;
             COM_emailNotification( $msgData );
     	} else {
-        	COM_errorLog("MG Upload: Error - Did not find any moderators to email");
+        	Log::write('system',Log::ERROR,"MG Upload: Error - Did not find any moderators to email");
     	}
         COM_updateSpeedlimit ('mgnotify');
     }