+++ public_html/mediagallery/album_rpc.php
@@ -1,35 +1,21 @@
 <?php
-// +--------------------------------------------------------------------------+
-// | Media Gallery Plugin - glFusion CMS                                      |
-// +--------------------------------------------------------------------------+
-// | album_rpc.php                                                            |
-// |                                                                          |
-// | AJAX component to retrieve album attributes                              |
-// +--------------------------------------------------------------------------+
-// | Copyright (C) 2009-2015 by the following authors:                        |
-// |                                                                          |
-// | Mark A. Howard         mark AT usable-web DOT com                        |
-// | Mark R. Evans          mark AT glfusion DOT org                          |
-// +--------------------------------------------------------------------------+
-// |                                                                          |
-// | This program is free software; you can redistribute it and/or            |
-// | modify it under the terms of the GNU General Public License              |
-// | as published by the Free Software Foundation; either version 2           |
-// | of the License, or (at your option) any later version.                   |
-// |                                                                          |
-// | This program is distributed in the hope that it will be useful,          |
-// | but WITHOUT ANY WARRANTY; without even the implied warranty of           |
-// | MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the            |
-// | GNU General Public License for more details.                             |
-// |                                                                          |
-// | You should have received a copy of the GNU General Public License        |
-// | along with this program; if not, write to the Free Software Foundation,  |
-// | Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.          |
-// |                                                                          |
-// +--------------------------------------------------------------------------+
+/**
+* glFusion CMS - Media Gallery Plugin
+*
+* AJAX component to retrieve album attributes
+*
+* @license GNU General Public License version 2 or later
+*     http://www.opensource.org/licenses/gpl-license.php
+*
+*  Copyright (C) 2002-2021 by the following authors:
+*   Mark R. Evans   mark AT glfusion DOT org
+*
+*/
 
 require_once '../lib-common.php';
 require_once $_CONF['path'].'plugins/mediagallery/include/init.php';
+
+use \glFusion\Log\Log;
 
 global $_CONF, $_MG_CONF, $MG_albums;
 
@@ -40,7 +26,7 @@ if ( COM_isAnonUser() && $_MG_CONF['loginrequired'] == 1 )  {
 }
 
 if( $_MG_CONF['verbose'] ) {
-    COM_errorLog( 'album_rpc.php: invocation ------------------------' );
+    Log::write('system',Log::DEBUG,'album_rpc.php: invocation *******');
 }
 
 if( isset( $_REQUEST['aid'] ) ) {
@@ -48,14 +34,14 @@ if( isset( $_REQUEST['aid'] ) ) {
     // retrieve the album_id passed
     $album_id = COM_applyFilter( $_REQUEST['aid'], true );
     if( $_MG_CONF['verbose'] ) {
-        COM_errorLog( 'album_id=' . $album_id );
+        Log::write('system',Log::DEBUG,'album_id=' . $album_id );
     }
 
     // initialize the $MG_albums array
     MG_initAlbums();
     $albums = sizeof( $MG_albums );
     if( $_MG_CONF['verbose'] ) {
-        COM_errorLog( 'initialized ' . $albums . ' albums' );
+        Log::write('system',Log::DEBUG,'initialized ' . $albums . ' albums' );
     }
 
     // check to ensure we have a valid album_id
@@ -66,7 +52,7 @@ if( isset( $_REQUEST['aid'] ) ) {
         // retrieve the valid filetypes
         $valid_types = MG_getValidFileTypes( $album_id );
     } else {
-        COM_errorLog( 'album_rpc.php: invalid album id' );
+        Log::write('system',Log::ERROR,'album_rpc.php: invalid album id' );
         $size_limit = 0;
         $valid_types = '';
     }
@@ -75,13 +61,13 @@ if( isset( $_REQUEST['aid'] ) ) {
     echo $size_limit . '%' . $valid_types;
 
     if( $_MG_CONF['verbose'] ) {
-        COM_errorLog( 'size_limit=' . $size_limit );
-        COM_errorLog( 'valid_types=' . $valid_types );
-        COM_errorLog( 'album_rpc.php: normal termination ----------------' );
+        Log::write('system',Log::DEBUG,'size_limit=' . $size_limit );
+        Log::write('system',Log::DEBUG,'valid_types=' . $valid_types );
+        Log::write('system',Log::DEBUG,'album_rpc.php: normal termination ----------------' );
     }
 
 } else {
-    COM_errorLog( 'album_rpc.php: invocation with no album parameter' );
+    Log::write('system',Log::ERROR,'album_rpc.php: invocation with no album parameter' );
 }
 exit(0);
 ?>