+++ public_html/mediagallery/fslideshow.php
@@ -1,4 +1,16 @@
 <?php
+/**
+* glFusion CMS - Media Gallery Plugin
+*
+* Flash Slideshow
+*
+* @license GNU General Public License version 2 or later
+*     http://www.opensource.org/licenses/gpl-license.php
+*
+*  Copyright (C) 2002-2021 by the following authors:
+*   Mark R. Evans   mark AT glfusion DOT org
+*
+*/
 // +--------------------------------------------------------------------------+
 // | Media Gallery Plugin - glFusion CMS                                      |
 // +--------------------------------------------------------------------------+